+++ HISTORY.md
@@ -4,6 +4,8 @@
 ## not yet released, version 3.1.1
 
 - Fixed #621: a bug in parsing implicit multiplications like `(2)(3)+4`.
+- Fixed #623: `nthRoot` of zero with a negative root returned `0` instead of
+  `Infinity`.
 
 
 ## 2016-03-19, version 3.1.0